requirements.txt
asgiref==3.3.1
boto3==1.16.25
botocore==1.19.25
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
defusedxml==0.6.0
Django==3.1.3
django-allauth==0.43.0
django-rest-auth==0.9.5
django-storages==1.10.1
djangorestframework==3.12.2
idna==2.10
jmespath==0.10.0
oauthlib==3.1.0
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
